requirements.txt
Django==3.0.6
ipdb==0.13.2
django-filer==1.7.1
python-slugify==4.0.0
